fixing cache

+++ server/src/external/stripe/handleStripeWebhookEvent.ts
@@ -1,12 +1,13 @@
-import type { AppEnv, Organization } from "@autumn/shared";
+import type { Organization } from "@autumn/shared";
 import chalk from "chalk";
 import { Stripe } from "stripe";
-import type { DrizzleCli } from "@/db/initDrizzle.js";
 import { createStripeCli } from "@/external/connect/createStripeCli.js";
 import { CusService } from "@/internal/customers/CusService.js";
 import { unsetOrgStripeKeys } from "@/internal/orgs/orgUtils.js";
 import type { ExtendedRequest } from "@/utils/models/Request.js";
+import type { AutumnContext } from "../../honoUtils/HonoEnv.js";
 import { deleteCachedApiCustomer } from "../../internal/customers/cusUtils/apiCusCacheUtils/deleteCachedApiCustomer.js";
+import { setCachedApiCusProducts } from "../../internal/customers/cusUtils/apiCusCacheUtils/setCachedApiCusProducts.js";
 import type { Logger } from "../logtail/logtailUtils.js";
 import { handleCheckoutSessionCompleted } from "./webhookHandlers/handleCheckoutCompleted.js";
 import { handleCusDiscountDeleted } from "./webhookHandlers/handleCusDiscountDeleted.js";
@@ -33,13 +34,10 @@ const logStripeWebhook = ({
 	);
 };
 
+const updateProductEvents = ["customer.subscription.updated"];
+
 const coreEvents = [
-	"customer.subscription.created",
-	"customer.subscription.updated",
 	"customer.subscription.deleted",
-	"invoice.paid",
-	"invoice.created",
-	"invoice.finalized",
 	"subscription_schedule.canceled",
 	"checkout.session.completed",
 ];
@@ -47,19 +45,17 @@ const coreEvents = [
 const handleStripeWebhookRefresh = async ({
 	eventType,
 	data,
-	db,
-	org,
-	env,
-	logger,
+	ctx,
 }: {
 	eventType: string;
 	data: any;
-	db: DrizzleCli;
-	org: Organization;
-	env: AppEnv;
-	logger: any;
+	ctx: AutumnContext;
 }) => {
-	if (coreEvents.includes(eventType)) {
+	const { db, logger, org, env } = ctx;
+	if (
+		coreEvents.includes(eventType) ||
+		updateProductEvents.includes(eventType)
+	) {
 		const stripeCusId = data.object.customer;
 		if (!stripeCusId) {
 			logger.warn(
@@ -86,17 +82,32 @@ const handleStripeWebhookRefresh = async ({
 			return;
 		}
 
-		// await deleteCusCache({
-		// 	db,
-		// 	customerId: cus.id!,
-		// 	org,
-		// 	env,
-		// });
-		await deleteCachedApiCustomer({
-			customerId: cus.id!,
-			orgId: org.id,
-			env,
-		});
+		if (updateProductEvents.includes(eventType)) {
+			const fullCus = await CusService.getFull({
+				db,
+				idOrInternalId: cus.id!,
+				orgId: org.id,
+				env,
+				withEntities: true,
+				withSubs: true,
+			});
+
+			await setCachedApiCusProducts({
+				ctx,
+				fullCus,
+				customerId: cus.id!,
+			});
+		} else {
+			await deleteCachedApiCustomer({
+				customerId: cus.id!,
+				orgId: org.id,
+				env,
+				source: {
+					stripeWebhook: true,
+					eventType,
+				},
+			});
+		}
 	}
 };
 
@@ -104,20 +115,23 @@ const handleStripeWebhookRefresh = async ({
  * Handles Stripe webhook events after org/env extraction
  */
 export const handleStripeWebhookEvent = async ({
+	ctx,
 	event,
-	db,
-	org,
-	env,
-	logger,
-	req,
+	// db,
+	// org,
+	// env,
+	// logger,
+	// req,
 }: {
+	ctx: AutumnContext;
 	event: Stripe.Event;
-	db: DrizzleCli;
-	org: Organization;
-	env: AppEnv;
-	logger: Logger;
-	req: ExtendedRequest;
+	// db: DrizzleCli;
+	// org: Organization;
+	// env: AppEnv;
+	// logger: Logger;
+	// req: ExtendedRequest;
 }) => {
+	const { db, logger, org, env } = ctx;
 	logStripeWebhook({ logger, org, event });
 
 	try {
@@ -136,7 +150,7 @@ export const handleStripeWebhookEvent = async ({
 			case "customer.subscription.updated": {
 				const subscription = event.data.object;
 				await handleSubscriptionUpdated({
-					req,
+					req: ctx as ExtendedRequest,
 					db,
 					org,
 					subscription,
@@ -149,7 +163,7 @@ export const handleStripeWebhookEvent = async ({
 
 			case "customer.subscription.deleted":
 				await handleSubDeleted({
-					req,
+					req: ctx as ExtendedRequest,
 					stripeCli,
 					data: event.data.object,
 					logger,
@@ -159,7 +173,7 @@ export const handleStripeWebhookEvent = async ({
 			case "checkout.session.completed": {
 				const checkoutSession = event.data.object;
 				await handleCheckoutSessionCompleted({
-					req,
+					req: ctx as ExtendedRequest,
 					db,
 					data: checkoutSession,
 					org,
@@ -177,7 +191,7 @@ export const handleStripeWebhookEvent = async ({
 					invoiceData: invoice,
 					env,
 					event,
-					req,
+					req: ctx as ExtendedRequest,
 				});
 				break;
 			}
@@ -187,7 +201,7 @@ export const handleStripeWebhookEvent = async ({
 					stripeCli,
 					env,
 					event,
-					req,
+					req: ctx as ExtendedRequest,
 				});
 				break;
 
@@ -234,7 +248,6 @@ export const handleStripeWebhookEvent = async ({
 					discount: event.data.object,
 					env,
 					logger,
-					res: req,
 				});
 				break;
 		}
@@ -272,10 +285,7 @@ export const handleStripeWebhookEvent = async ({
 		await handleStripeWebhookRefresh({
 			eventType: event.type,
 			data: event.data,
-			db,
-			org,
-			env,
-			logger,
+			ctx,
 		});
 	} catch (error) {
 		logger.error(`Stripe webhook, error refreshing cache!`, { error });
